Better fix for jdk11-compatibility of Process.pidValue
