+++ pom.xml
@@ -12,9 +12,6 @@
         <selenium.version>3.12.0</selenium.version>
         <kotlin.version>1.3.71</kotlin.version>
         <exec.mainClass>org.jitsi.jibri.MainKt</exec.mainClass>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <kotlin.compiler.jvmTarget>1.8</kotlin.compiler.jvmTarget>
         <smack.version>4.2.4-47d17fc</smack.version>
         <jersey.version>2.30.1</jersey.version>
         <jackson.version>2.9.5</jackson.version>
@@ -24,6 +21,33 @@
         <ktor.version>1.3.1</ktor.version>
         <jicoco.version>1.1-54-g5387346</jicoco.version>
     </properties>
+
+    <profiles>
+        <profile>
+            <id>target-jdk8</id>
+            <activation>
+                <jdk>1.8</jdk>
+            </activation>
+            <properties>
+                <maven.compiler.source>1.8</maven.compiler.source>
+                <maven.compiler.target>1.8</maven.compiler.target>
+                <kotlin.compiler.jvmTarget>1.8</kotlin.compiler.jvmTarget>
+                <kcond>jdk8</kcond>
+            </properties>
+        </profile>
+        <profile>
+            <id>target-jdk9</id>
+            <activation>
+                <jdk>[9,)</jdk>
+            </activation>
+            <properties>
+                <maven.compiler.source>9</maven.compiler.source>
+                <maven.compiler.target>9</maven.compiler.target>
+                <kotlin.compiler.jvmTarget>9</kotlin.compiler.jvmTarget>
+                <kcond>jdk9</kcond>
+            </properties>
+        </profile>
+    </profiles>
 
     <dependencies>
         <dependency>
@@ -212,6 +236,7 @@
                         <configuration>
                             <sourceDirs>
                                 <sourceDir>${project.basedir}/src/main/kotlin</sourceDir>
+                                <sourceDir>${project.basedir}/src/main/kotlin-${kcond}</sourceDir>
                                 <sourceDir>${project.basedir}/src/main/java</sourceDir>
                             </sourceDirs>
                             <args>
